% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man_expected_files.R
\name{expected_files}
\alias{expected_files}
\title{Files expected by \code{tsv2label}}
\description{
For a thorough introduction to the package, including what the dictionary
files should actually look like and how they should be stored, see
\href{https://github.com/DesiQuintans/tsv2label}{\verb{the online readme on Github}}.
}
