% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_factorise_with_dictionary.R
\name{factorise_with_dictionary}
\alias{factorise_with_dictionary}
\alias{factorize_with_dictionary}
\title{Convert variables to Factors with a data dictionary}
\usage{
factorise_with_dictionary(df, path)

factorize_with_dictionary(df, path)
}
\arguments{
\item{df}{(Dataframe) A dataframe to label.}

\item{path}{(Character) Path to the dataset's dictionary files, which is
either a folder or a .zip file. See \link{expected_files} for more info.}
}
\value{
Mutates \code{df} in-place in the global environment.
}
\description{
Datasets imported into R from foreign analysis packages (SPSS, SAS, MATLAB)
often come with "value labels" which, for example, tell you that a \code{0} in a
column actually means \code{"Travelling locally"} or something. R represents such
categorical data as Factors. Some other R packages put these labels into an
attribute called \code{"labels"} (note plural) as a temporary measure until you
decide to convert them into factors. \emph{This} package goes straight to
converting them so that they're ready to use.
}
\examples{
\donttest{
# An unlabelled dataset that is built-into this package.
my_poker <- poker

# system.file() gives the path to a dictionary for this dataset that is also
# built-into this package.
factorise_with_dictionary(my_poker, system.file("extdata/poker", package = "tsv2label"))

##  (1/4)  Using 'values_flip' for COIN FLIP.
##  (2/4)  Using 'values_hands' for CLASS.
##  (3/4)  Using 'values_ranks' for C1, C2, C3, C4, C5.
##  (4/4)  Using 'values_suits' for S1, S2, S3, S4, S5.
##
##  Peeking at 'levels(my_poker[["COIN FLIP"]])', built from
##  'values coin flip':
##  Heads, Tails
##
##  Peeking at 'levels(my_poker[["CLASS"]])', built from
##  'values_hands':
##  Nothing in hand, One pair, Two pairs, Three of a kind,
##  Straight, Flush, Full house, Four of a kind, Straight
##  flush, Royal flush
##
##  Peeking at 'levels(my_poker[["C1"]])', built from
##  'values_ranks':
##  Ace, 2, 3, 4, 5, 6, 7, 8, 9, 10, Jack, Queen, King
##
##  Peeking at 'levels(my_poker[["S1"]])', built from
##  'values_suits':
##  Hearts, Spades, Diamonds, Clubs
}
}
