% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man_poker_dataset.R
\docType{data}
\name{poker}
\alias{poker}
\title{Poker Hand example dataset}
\format{
\subsection{\code{poker}}{

A data frame with 100 rows and 17 columns.
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Poker+Hand}
}
\usage{
poker
}
\description{
This is the first 100 rows of the training subset of the "Poker Hand"
dataset, originally from the UCI Machine Learning Repository. This package
comes with a set of labels for it, located in the \code{extdata/poker} folder
where this package is installed --- use \code{find.package("tsv2label")} to find
out where that is.
}
\details{
I chose it as example data for \code{tsv2label} because it codes a playing card's
suit and rank as integers that need to be relabelled with their actual
meanings. It shows features of \code{tsv2label} like:
\enumerate{
\item One \code{factor_file} can be applied to many columns (e.g. all 5 cards'
ranks are named using the same \code{values_ranks} file).
\item A \code{factor_file} can also be dedicated to one column only (e.g. \code{value_hands}
factorises the \code{CLASS} column only).
\item Factor conversion is not performed on columns without a named \code{factor_file}
(e.g. \code{ID}, \code{CAT}).
\item Columns are not renamed if they have no \code{rename} value (e.g. \code{CAT}).
\item A data dictionary can contain columns that don't exist in the data (e.g. \code{not_exist}).
\item A data dictionary can omit columns that \strong{do} exist in the data (e.g. \code{LIST_COL}).
\item Non-syntactic names are allowed (e.g. \verb{COIN FLIP}).
\item List columns are supported (e.g. \code{LIST_COL}).
\item Conversion of truthy and falsy variables (\code{LOGICAL_CHAR} and \code{LOGICAL_INT})
using built-in converters.
}
}
\keyword{datasets}
