% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man_tsv2label.R
\name{tsv2label}
\alias{tsv2label}
\title{Label, describe, rename, and factorise datasets using a data dictionary}
\description{
This package lets you describe and label datasets from human-readable data
dictionaries in the form of tab-separated spreadsheets (\code{.TSV} or \code{.TXT}).
This format is ideal because it can be hand-edited in Excel; can be built
by copying and pasting from existing messy data dictionaries; can be cleaned
and reshaped with regular expressions; and can be committed, tracked, and
diffed in version control software. \code{tsv2label} supports variable labels,
variable renaming, and value labelling (through automatic conversion to
Factor). It modifies dataframes in-place, making it suitable for large data.
}
\details{
For a thorough introduction to the package, including what the dictionary
files should actually look like and how they should be stored, see
\href{https://github.com/DesiQuintans/tsv2label}{\verb{the online readme on Github}}..
}
