% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_describe_with_dictionary.R
\name{describe_with_dictionary}
\alias{describe_with_dictionary}
\title{Add descriptions to variables using a data dictionary}
\usage{
describe_with_dictionary(df, path, quiet = FALSE, up = 2)
}
\arguments{
\item{df}{(Dataframe) A dataframe to label.}

\item{path}{(Character) Path to the dataset's dictionary files, which is
either a folder or a .zip file. See \link{expected_files} for more info.}

\item{up}{(Integer) The number of environments to step back in, determining
where this function will be evaluated. See \code{\link[=parent.frame]{parent.frame()}} for details.
The default value of \code{up = 2} is usually appropriate.}
}
\value{
Mutates \code{df} in-place in the specified environment.
}
\description{
Variable descriptions (the attribute called \code{"label"}) are used by lots of
R packages to add functionality. For example, RStudio can display labels in
\code{\link[=View]{View()}}, the \code{gtsummary} package uses the label attribute to name
variables in its output tables where possible, and the \code{siftr} package allows
you to search the labels to find variables in large datasets.
}
\examples{
\donttest{
# An unlabelled dataset that is built-into this package.
my_poker <- poker

# system.file() gives the path to a dictionary for this dataset, which is also
# built-into this package.
describe_with_dictionary(my_poker, system.file("extdata/poker", package = "tsv2label"))

##  head() of 'label' attribute:
##  ID Some random integers.
##  S1 Suit of card #1
##  C1 Rank of card #1
##  S2 Suit of card #2
##  C2 Rank of card #2
##  S3 Suit of card #3
}
}
