% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man_poker_dataset.R
\docType{data}
\name{poker}
\alias{poker}
\title{Poker Hand example dataset}
\format{
\subsection{\code{poker}}{

A data frame with 100 rows and 15 columns.
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Poker+Hand}
}
\usage{
poker
}
\description{
This is the first 100 rows of the training subset of the "Poker Hand"
dataset, originally from the UCI Machine Learning Repository. This package
comes with a set of labels for it, located in the \code{extdata/poker} folder
where this package is installed --- use \code{find.package("tsv2label")} to find
out where that is.
}
\details{
I chose it as example data for \code{tsv2label} because it codes a playing card's
suit and rank as integers that need to be relabelled with their actual
meanings. It shows how the same \code{value_file} can be used to add value label
to many columns (e.g. labelling the 5 card ranks columns using the same
file), or how a \code{value_file} can be dedicated to one column only (e.g.
\code{value_hands} to label the \code{CLASS} column). I have added 2 extra columns of
different types (\code{ID}, \code{CAT}) to show that \code{tsv2label} leaves
them alone if no \code{value_file} is specified for them in \code{index.tsv}. I have
also added a column with a non-syntactic name (\verb{COIN FLIP}) and a list column
(\code{LIST_COL}) for testing.
}
\keyword{datasets}
